refactor(connector): rename connectors to connector
