+++ pkg/cmd/connector/namespaces/create/create.go
