+++ pkg/cmd/connector/namespaces/namespaces.go
@@ -1,8 +1,8 @@
 package namespaces
 
 import (
-	"github.com/redhat-developer/app-services-cli/pkg/cmd/connectors/namespaces/create"
-	"github.com/redhat-developer/app-services-cli/pkg/cmd/connectors/namespaces/list"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/connector/namespaces/create"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/connector/namespaces/list"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/factory"
 	"github.com/spf13/cobra"
 )
